refactor(dafny): extract proof lemma library into dafny/lemmas
Move the shape-recognising proof strategies (concat-fold bridges,
additive-monoid op lemmas, rev anti-homomorphism + its step bridges) out of the
~2.2k-line toplevel.rs into a dedicated src/codegen/dafny/lemmas.rs behind two
registry aggregators: `algebra_lemmas` (proved helper lemmas + forall-lifted
facts) and `list_bridges` (the induction-step cons-decomposition asserts).
`emit_verify_law` now calls the two aggregators instead of inlining each
strategy, so adding a new lemma family is a local edit in one module.

Pure mechanical refactor, NO behaviour change: the emitted Dafny is
BYTE-IDENTICAL across the whole proof-corpus (verified by SHA of all 110 entry
.dfy before/after; proof_spec 89/89). The recognisers (detect_additive_op,
detect_rev_fn, collect_concat_bridges) are pure AST walks and already
backend-neutral — this is the seam for a backend-neutral recogniser/renderer
split (a Lean renderer of the List-only families would kernel-certify what Z3
proves today).
